Created folder structure and reorganized project
