+++ __tests__/App.js
@@ -1,6 +1,6 @@
 import 'react-native';
 import React from 'react';
-import App from '../App';
+import App from '../app/screens/App';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
