+++ app/components/Navbar.js
@@ -19,35 +19,35 @@ export default class Navbar extends Component<{}> {
                 <View style={styles.button} >
                     <TouchableHighlight onPress={() => this.props.navigation.navigate('HomePage')}>
                         <Image
-                            source={require('./images/home.png')}
+                            source={require('../../images/home.png')}
                         />
                     </TouchableHighlight>
                 </View>
                 <View style={styles.button} >
                     <TouchableHighlight onPress={() => this.props.navigation.navigate('ExperimentBrowse')}>
                         <Image
-                            source={require('./images/experiments.png')}
+                            source={require('../../images/experiments.png')}
                         />
                     </TouchableHighlight>
                 </View>
                 <View style={styles.biggerButton} >
                     <TouchableHighlight onPress={this._onPressButton}>
                         <Image
-                            source={require('./images/completeSurvey.png')}
+                            source={require('../../images/completeSurvey.png')}
                         />
                     </TouchableHighlight>
                 </View>
                 <View style={styles.button} >
                     <TouchableHighlight onPress={this._onPressButton}>
                         <Image
-                            source={require('./images/stats.png')}
+                            source={require('../../images/stats.png')}
                         />
                     </TouchableHighlight>
                 </View>
                 <View style={styles.button} >
                     <TouchableHighlight onPress={this._onPressButton}>
                         <Image
-                            source={require('./images/userProfile.png')}
+                            source={require('../../images/userProfile.png')}
                         />
                     </TouchableHighlight>
                 </View>
