+++ app/screens/App.js
@@ -12,7 +12,7 @@ import SignUp from './SignUp.js'
 import ExperimentBrowse from './ExperimentBrowse.js'
 import HomePage from './HomePage.js'
 import { AppRegistry, ScrollView, View} from 'react-native';
-import Navbar from './Navbar.js'
+import Navbar from '../components/Navbar.js'
 
 export default class App extends Component<{}> {
