+++ app/screens/SignUp.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import {Card} from 'react-native-elements';
-import LoginForm from './LoginForm'
+import LoginForm from '../components/LoginForm'
 
 
 export default class SignUp extends Component<{}> {
