+++ index.js
@@ -1,4 +1,4 @@
 import { AppRegistry } from 'react-native';
-import App from './App';
+import App from './app/screens/App';
 
 AppRegistry.registerComponent('LabsApp', () => App);
